% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_snow.R
\name{aggregate_snow}
\alias{aggregate_snow}
\title{Aggregates a binary snow mask to a fractional snow cover raster}
\usage{
aggregate_snow(x, fact)
}
\arguments{
\item{x}{SpatRaster - Binary snow mask (1 == snow, 0 == no snow)}

\item{fact}{Integer - Aggregation factor (number of pixels in each direction)}
}
\value{
Continuous SpatRaster with snow cover fraction (0-1) per aggregated cell
}
\description{
Aggregates a binary snow mask to a fractional snow cover raster
}
\examples{
fsc <- aggregate_snow(x, fact=10)
}
