% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/despeckle.R
\name{despeckle_snow}
\alias{despeckle_snow}
\title{Despeckles snow patches based on a binary snow mask (1==snow, 0==no snow)}
\usage{
despeckle_snow(
  x,
  window = 3,
  threshold = NULL,
  max_distance = NULL,
  min_pixels = NULL
)
}
\arguments{
\item{x}{SpatRaster - Binary snow mask (1 == snow, 0 == no snow)}

\item{window}{Integer - Odd-Sized neighborhood window for majority filtering}

\item{threshold}{Integer or NULL - Neighborhood threshold.  If NULL, strict majority is used.}

\item{max_distance}{Integer - Maximum allowed distance to other snow pixles (map units)}

\item{min_pixels}{Integer - Minimum numbers of connected pixels required for snow patch to be retained.}
}
\value{
SpatRaster - Despeckled binary snow mask
}
\description{
Despeckles snow patches based on a binary snow mask (1==snow, 0==no snow)
}
\examples{
#Basic majority-based despeckling
snow_clean <- despeckle_snow(
x,
window = 3
)

#Remove all isolated snowpatches
snow_clean <- despeckle_snow(
x,
window=3,
min_pixels=10
)

#Remove spatially isolated snow pixels
snow_clean <- despeckle_snow(
x,
window=3,
max_distance=40
)
}
