% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_snow.R
\name{detect_snow}
\alias{detect_snow}
\title{Detect Snow using Spectral Indices}
\usage{
detect_snow(
  x,
  index = "ndsi",
  bands = list(green = 2, swir = 4, red = 1, blue = 3),
  threshold = NULL
)
}
\arguments{
\item{x}{SpatRaster - A tif containing multiple bands from a multispectral image}

\item{index}{Character - The Index to use (currently supports "ndsi" and "rgb_brightness")}

\item{bands}{Named list with band indices for green, swir, red, and blue bands}

\item{threshold}{Numeric or NULL Threshold value for snow detection. If NULL, Otsu's method is used to determine the threshold.}
}
\value{
A list with three elements: the indexed Tif (SpatRaster), the binary snow mask (SpatRaster), and the threshold value(numeric).
}
\description{
Detect Snow using Spectral Indices
}
\examples{
detect_snow(x, index="ndsi", bands=list(green=2,swir=4),threshold=NULL)

}
