% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsc.R
\name{snow_cover_fraction}
\alias{snow_cover_fraction}
\title{Estimate Snow Cover Fraction (SCF) from NDSI}
\usage{
snow_cover_fraction(x, bands, method)
}
\arguments{
\item{x}{Spatraster - Multispectral image}

\item{bands}{Named list with "green" and "swir" band indices}

\item{method}{Character - Either "salomonson" (Salomonson & Appel, 2004) or "sigmoid" (Gascoin et al., 2020)}
}
\value{
Spatraster with Snow Cover Fraction values (clipped to 0-1)
}
\description{
Estimate Snow Cover Fraction (SCF) from NDSI
}
\examples{
scf <- snow_cover_fraction(x, bands=list(green=3, swir=10), method="salomonson")
}
